Added code for firestore cloud functions
Added code for firestore cloud functions
1. Create user
{
email:
password:
fname:
lname:
balance:
balanceonhold:
}
The balance on hold is what I thought might work for when a user places a bid it will move money from balance to balance on hold.

2. addBalance
will add money to the balance of the user whose id is passed

3. getProfile
to get the user's profile

4. postItem
{
owner:
item:
currentbid:
bidholder:
minfinalbid:
previousbids:
}
The previousbids would hold currentbid and bidholder so we could cycle back through the list/array if the bid is cancelled.